ref:#35 update comments
