+++ back/.babelrc
@@ -19,8 +19,6 @@
       {
         "root": ["."],
         "alias": {
-          // "common": "./src/common",
-          // "common-app": "./src/common-app",
           "core": "./src/core",
           "dals": "./src/dals",
           "pods": "./src/pods"
