+++ back/.env-sample
@@ -1,0 +1,5 @@
+NODE_ENV=development
+PORT=8081
+API_MOCK=false
+MONGO_URL=mongodb://localhost:27017/planning-poker
+API_URL=/api
